train/requirements.txt
anyconfig==0.14.0
blosc==1.11.3
lmdb==1.6.2
numpy
opencv-python==4.11.0.86
pillow==11.2.1
pymupdf
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
focal-loss-torch
onnx
onnxruntime
onnxruntime-extensions
transformers
matplotlib
nicegui
torch-geometric
tensorboard